backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
motor==3.6.0
pymongo==4.9.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
boto3==1.28.78
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
